+++ Fractals.jl/src/interactive.jl
@@ -148,7 +148,7 @@ function prompt_ifs_and_render(path::AbstractString;
 
     final_outpath = _normalize_media_outpath(outpath)
     img = make_image(ifs; resolution=(height, width))
-    save(final_outpath, img)
+    _save_image_with_source_limits(final_outpath, img, ifs.limits)
     println("Saved image to $(final_outpath)")
     return ifs
 end